Fix compile warnings, cast size_t to long unsigned int for printf

+++ src/stratagus/groups.cpp
@@ -127,7 +127,7 @@ void SaveGroups(CFile &file)
 	file.printf("--- MODULE: groups\n\n");
 
 	for (int g = 0; g < NUM_GROUPS; ++g) {
-		file.printf("Group(%d, %d, {", g, Groups[g].getUnits().size());
+		file.printf("Group(%d, %lu, {", g, (long unsigned int)Groups[g].getUnits().size());
 		for (size_t i = 0; i < Groups[g].getUnits().size(); ++i) {
 			file.printf("\"%s\", ", UnitReference(*Groups[g].getUnits()[i]).c_str());
 		}
